Extract metadata related to GeoJSONMetadataBuilder
